remove header button into own component folder
